+++ config/config.go
@@ -91,8 +91,11 @@ var (
 	// BuildImage is the Trident image that was built
 	BuildImage = "netapp/trident:" + orchestratorVersion + "-custom.0"
 
-	// EtcdImage is the etcd image that Trident should be deployed with
-	BuildEtcdImage = "quay.io/coreos/etcd:v3.2.19"
+	// BuildEtcdVersion is the etcd version that Trident should be deployed with
+	BuildEtcdVersion = "v3.2.19"
+
+	// BuildEtcdImage is the etcd image that Trident should be deployed with
+	BuildEtcdImage = "quay.io/coreos/etcd:" + BuildEtcdVersion
 
 	OrchestratorVersion = utils.MustParseDate(version())