+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - **Kubernetes:** Updated etcd to v3.2.19 and client-go to v7.0.0.
 - **Kubernetes:** Added --previous switch to 'tridentctl logs' command.
 - **Kubernetes:** Added liveness probes to the trident-main and etcd containers.
+- **Kubernetes:** Added --trident-image and --etcd-image switches to 'tridentctl install' command.
 
 ## v18.04.0